chore: Organização do codigo (BoardCheckExtensions)
